FixAndRefactor : debateboard,reply,applicant can work at swagger-ui
위 3가지 entity의 id를 int로 바꾸고 swagger를 전부 돌려보고 문제가 있는 요소들을 찾아 고쳐 스웨거에서 정상 동작하는 것을 확인했습니다.

+++ backend/src/main/java/com/ssafy/backend/api/request/applicantState/ApplicantStateRegiPostReq.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class ApplicantStateRegiPostReq {
     int side;
     int order;
-    long board_no;
+    int board_no;
 }
