+++ backend/src/main/java/com/ssafy/backend/api/request/reply/ReplyChangeReq.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ReplyChangeReq {
-    long reply_no;
+    int reply_no;
     String context;
 }
